[interp] fix warning
```
interp/interp.c: In function ‘ves_real_abort’:
interp/interp.c:251:17: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 2 has type ‘int’ [-Wformat=]
   printf ("\t[%ld] 0x%08x %0.5f\n", sp-stack, sp[-1].data.i, sp[-1].data.f);
               ~~^                   ~~~~~~~~
               %d
```

+++ mono/mini/interp/interp.c
@@ -255,7 +255,7 @@ ves_real_abort (int line, MonoMethod *mh,
 	g_print ("0x%04x %02x\n", ip-(const unsigned short *) header->code, *ip);
 	mono_metadata_free_mh (header);
 	if (sp > stack)
-		printf ("\t[%ld] 0x%08x %0.5f\n", sp-stack, sp[-1].data.i, sp[-1].data.f);
+		printf ("\t[%lld] 0x%08x %0.5f\n", (gint64) (sp - stack), sp[-1].data.i, sp[-1].data.f);
 }
 
 #define ves_abort() \
